Models

ENAModel

Main.EpistemicNetworkAnalysis.ENAModel — Type

ENAModel(
    # Required
    data::DataFrame,
    codes::Array{Symbol,1},
    conversations::Array{Symbol,1},
    units::Array{Symbol,1};

    # Optional
    rotation::AbstractLinearENARotation=SVDRotation(),
    unitFilter::Function=unit->true,
    edgeFilter::Function=edge->edge.kind == :undirected,
    windowSize::Real=Inf,
    sphereNormalize::Bool=true,
    dropEmpty::Bool=false,
    recenterEmpty::Bool=false
)

Construct an undirected ENA model. Nodes are positioned to maximize goodness of fit between plotted points and units' weighted average of edge midpoints.

Arguments

The minimum required arguments are:

data: DataFrame containing your qualitative data. Each row should represent one "line" or the smallest codable unit. Columns should include metadata information, the "text" or the qualitative unit itself, and binary-coded qualitative codes.

codes: Array listing names of columns to use as the qualitative codes in the model

conversation: As above, but for columns to use to distinguish "conversations," segments the model should not count connections between

units: As above, but for columns to use to distinguish "speakers" or your units of analysis

The common optional arguments are:

rotateBy: "Rotation" or dimension reduction the model should use for determining axes for downstream plotting and analysis

windowSize: Size of the sliding window for counting connections, where a window size of 1 counts only connections occuring on the same line

dropEmpty: Whether the model should drop units of analysis that accumulated no connections

Finally, the less common but occassionally useful optional arguments are:

unitFilter: Function for deciding which units of analysis to include in the model, such as based on its metadata

edgeFilter: As above, but for which edges to include in the model, such as based on which codes it connects. Ensure that only :undirected edges are included

sphereNormalize: Whether the model should normalize units of analysis so that, for example, speakers who talk more are still considered similar to those who talk less but about the same subjects

recenterEmpty: Whether the model should move empty units of analysis to the mean of all units, instead of leaving them at the zero origin

Fields

Once the model is constructed, it will have the following fields:

data, codes, conversations, units, and rotation: Copies of the argument values given above

metadata: DataFrame of original non-code columns, each row corresponding to one unit of analysis. A unitID column is added to represent each unit's unique identifier

points: DataFrame, where rows correspond to plotted point dimensions and columns correspond to units of analysis

pointsHat: As above, but for approximate dimensions used for measuring model goodness of fit

pointsNodes: As above, but for the optimized node positions along each plotted dimension

accum: DataFrame, where rows correspond to units of analysis and columns correspond to edges, counting the (normalized) number of connections that unit accumulated for that edge

accumHat: As above, but for the approximated counts, used for measuring model goodness of fit

edges: DataFrame, where rows correspond to edges, with the following columns: edgeID, kind, ground, and response

nodes: DataFrame, where rows correspond to nodes and columns correspond to approximated counts for each edge, used for computing pointsNodes

embedding: DataFrame, where rows correspond to plotted point dimensions and columns correspond edges. Used for computing points, pointsHat, and pointsNodes. Additional columns may be added depending on the model's rotation, corresponding to statistical tests for each dimension

config: NamedTuple, storing additional configuration options, such as unitFilter, sphereNormalize, and so on

Example

using EpistemicNetworkAnalysis

# Load example dataset
data = loadExample("shakespeare")

# Base settings
codes = [:Love, :Death, :Honor, :Men, :Women]
conversations = [:Play, :Act, :Scene]
units = [:Play, :Speaker]

# Rotation settings
rotation = TopicRotation(
    "Women-Death vs. Honor",
    [:Women, :Death],
    [:Honor]
)

# Construct model
model = ENAModel(
    data, codes, conversations, units,
    windowSize=4,
    rotateBy=rotation,
    dropEmpty=false
)

# Display model overview
show(model)

# Display summary statistics
show(summary(model))

# Display and save plot
p = plot(model)
show(p)
savefig(p, "example.svg")

# Save model for later use in Julia
serialize("example.ena", model)

# Save model for easy sharing with collaborators
to_xlsx("example.xlsx", model)

DigraphENAModel

Main.EpistemicNetworkAnalysis.DigraphENAModel — Type

DigraphENAModel(
    # Required
    data::DataFrame,
    codes::Array{Symbol,1},
    conversations::Array{Symbol,1},
    units::Array{Symbol,1};

    # Optional
    rotation::AbstractLinearENARotation=SVDRotation(),
    unitFilter::Function=unit->true,
    edgeFilter::Function=edge->edge.kind == :directed,
    windowSize::Real=Inf,
    sphereNormalize::Bool=true,
    dropEmpty::Bool=false,
    recenterEmpty::Bool=false
)

Construct a directed ENA model. Nodes are positioned to maximize goodness of fit between plotted points and units' weighted average of edge vectors.

DigraphENAModel follows the same argument and field structure as ENAModel.

Ensure that edgeFilter only includes :directed edges.

BiplotENAModel

Main.EpistemicNetworkAnalysis.BiplotENAModel — Type

BiplotENAModel(
    # Required
    data::DataFrame,
    codes::Array{Symbol,1},
    conversations::Array{Symbol,1},
    units::Array{Symbol,1};

    # Optional
    rotation::AbstractLinearENARotation=SVDRotation(),
    unitFilter::Function=unit->true,
    sphereNormalize::Bool=true,
    dropEmpty::Bool=false,
    recenterEmpty::Bool=false
)

Construct a biplot model of unit-wise counts of code occurences, without measuring connections between codes. Model will have perfect goodness of fit between points and pointsHat, will be much simpler than other model types, but will lose most information compared to other model types.

BiplotENAModel follows the same argument and field structure as ENAModel, except edgeFilter and windowSize are in effect ignored.

Rotations

SVDRotation

Main.EpistemicNetworkAnalysis.SVDRotation — Type

SVDRotation()

The default rotation for linear ENA models. Reduces dimensions usings singular value decomposition

MeansRotation

Main.EpistemicNetworkAnalysis.MeansRotation — Type

MeansRotation(
    # Required, groups to compare along the x-axis
    groupVar1::Symbol,
    controlGroup1::Any,
    treatmentGroup1::Any,

    # Optional, groups to compare along subsequent axes
    args...;

    # Optional, whether to moderate the interactions between group dimensions
    moderated=false
)

Define a rotation for comparing pairs of groups, by maximizing the variance between pairs

See also: LDARotation and MulticlassRotation

Example

rotation = MeansRotation(
    :Play, "Romeo and Juliet", "Hamlet",
    :Act, 1, 5
)

Statistical Tests

Models using a MeansRotation will run the following statistical tests:

KruskalWallisTest for each dimension with a group pair

LDARotation

Main.EpistemicNetworkAnalysis.LDARotation — Type

LDARotation(
    groupVar::Symbol
)

Define a rotation for comparing multiple groups, via linear discriminant analysis

See also: MeansRotation and MulticlassRotation

Example

rotation = LDARotation(:Act)

Statistical Tests

Models using an LDARotation will run the following statistical tests:

KruskalWallisTest for each dimension

MulticlassRotation

Main.EpistemicNetworkAnalysis.MulticlassRotation — Type

MulticlassRotation(
    groupVar::Symbol
)

Define a rotation for comparing multiple groups, by maximizing between-group variance

See also: MeansRotation and LDARotation

Example

rotation = MulticlassRotation(:Act)

Statistical Tests

Models using an MulticlassRotation will run the following statistical tests:

KruskalWallisTest for each dimension

FormulaRotation

Main.EpistemicNetworkAnalysis.FormulaRotation — Type

FormulaRotation(
    regression_model1::Type{T},
    formula1::FormulaTerm,
    coef_index1::Int,
    contrast1::Union{Nothing,Dict},
    args...
) where {T <: RegressionModel}

Define a rotation that uses regression models to determine axes most closely associated with some linear trend

Note: RegressionModels must be imported from other stats packages

Note: contrasts are used to model categorical data

Example

using GLM

rotation = EpistemicNetworkAnalysis.FormulaRotation(
    LinearModel, @formula(edge ~ 1 + FinalGrade), 2, nothing
)

This will fit the x-axis to the FinalGrade metadata, because:

We use LinearModel from the GLM package

We use the formula edge ~ 1 + FinalGrade

And we use the 2nd coefficient of the LinearModel (in this case FinalGrade) to determine the values of the embedding

We have no categorical data in the LinearModel, so we leave the contrasts as nothing

Additional formulae may be used to define subsequent axes:

rotation = EpistemicNetworkAnalysis.FormulaRotation(
    LinearModel, @formula(edge ~ 1 + PretestGrade + PosttestGrade), 2, nothing,
    LinearModel, @formula(edge ~ 1 + PretestGrade + PosttestGrade), 3, nothing
)

Note: When multiple formulae are given, FormulaRotation finds the plane of the effects in the accum space, rotating it such that the first formula aligns with the x-axis, and the second formula aligns approximately with the y-axis. Unless this approximation is strong, a warning will be raised describing possible issues.

Statistical Tests

Models using a MeansRotation will run the following statistical tests:

R^2 and adjusted-R^2 for each dimension with a formula

TopicRotation

Main.EpistemicNetworkAnalysis.TopicRotation — Type

TopicRotation(
    topicName::AbstractString,
    controlNodes::Array{Symbol},
    treatmentNodes::Array{Symbol}
)

Define a rotation that places its x-axis through the mean of controlNodes on the left and the mean of treatmentNodes on the right, ie., through an a priori defined topic

Example

rotation = TopicRotation(
    "Women-Death vs. Honor",
    [:Women, :Death],
    [:Honor]
)

TrainedRotation

Main.EpistemicNetworkAnalysis.TrainedRotation — Type

TrainedRotation(
    trainmodel::AbstractLinearENAModel
)

Define a rotation that copies the embedding of an existing "training" model for use in a second "testing" model

Example

trainmodel = ENAModel(traindata, codes, conversations, units, rotation=LDARotation(:Act))
testmodel = ENAModel(testdata, codes, conversations, units, rotation=TrainedRotation(trainmodel))

Statistical Tests

Models using a TrainedRotation will run the same statistical tests as the trainmodel.