Helpful Functions

The following documentation describes helpful utility functions included with this package.

Once you're familiar with them, get involved!

Make a discussion post introducing yourself and sharing how you're using Epistemic Network Analysis

File an issue anytime you encounter a bug or are unable to make the package do what you need

loadExample

Main.EpistemicNetworkAnalysis.loadExample — Function

loadExample(name::AbstractString)

Load an example dataset as a DataFrame

Datasets

loadExample("shakespeare"): Loads the Shakespeare dataset, containing data on two plays, "Hamlet" and "Romeo and Juliet"

loadExample("transition"): Loads the Telling Stories of Transitions dataset, containing metadata and codes only, due to the sensitive nature of the underlying text

loadExample("toy"): Loads a minimal toy example, reproduced below

Group,Convo,Unit,Line,A,B,C
Red,1,X,1,0,0,1
Red,1,Y,2,1,0,0
Blue,1,Z,3,0,1,1
Blue,1,W,4,0,0,0
Red,1,X,5,0,0,1
Red,2,X,1,1,0,0
Red,2,Y,2,1,0,0
Blue,2,Z,3,0,1,1
Blue,2,W,4,0,0,0
Red,2,X,5,1,0,0

Loading Your Own Data

To load your own datasets, use DataFrame and CSV.File, which requires the DataFrames and CSV packages

using Pkg
Pkg.add("DataFrames")
using DataFrames

Pkg.add("CSV")
using CSV

data = DataFrame(CSV.File("filename_here.csv"))

deriveAnyCode!

Main.EpistemicNetworkAnalysis.deriveAnyCode! — Function

deriveAnyCode!(
    data::DataFrame,
    newColumnName::Symbol,
    oldColumnNames...
)

Add a new code column to data, derived from existing codes. The new code will be marked present where any of the old codes are present

Example

deriveAnyCode!(data, :Food, :Hamburgers, :Salads, :Cereal)

deriveAllCode!

Main.EpistemicNetworkAnalysis.deriveAllCode! — Function

deriveAllCode!(
    data::DataFrame,
    newColumnName::Symbol,
    oldColumnNames...
)

Add a new code column to data, derived from existing codes. The new code will be marked present only where all of the old codes are present on the same line

Example

deriveAllCode!(data, :ObservingStudentsLearning, :Observing, :Students, :Learning)

to_xlsx

Main.EpistemicNetworkAnalysis.to_xlsx — Function

to_xlsx(filename::AbstractString, model::AbstractENAModel)

Save a model to the disk as an Excel spreadsheet, useful for sharing results with others

See also serialize for saving models in a more efficient format that can be reloaded into Julia using deserialize

Note: a from_xlsx function does not exist, but is planned. The difficulty is that Excel data is at root a human-readable string format, and some components of some models are difficult to represent reliably as human-readable strings